+++ src/Inferno/Game.Player.h
@@ -124,7 +124,7 @@ namespace Inferno {
 
         void AutoselectPrimary();
         void AutoselectSecondary();
-        void GiveExtraLife(int lives = 1);
+        void GiveExtraLife(uint8 lives = 1);
 
     private:
         SoundUID _afterburnerSoundSig = 0;
